requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
googlemaps==4.10.0
google-cloud-secret-manager==2.21.0
google-cloud-aiplatform==1.69.0
pydantic==2.9.2
